refactor(tools): replace heavy git_intent parser with lightweight bash git semantics
